% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondiff.R
\docType{package}
\name{jsondiff-package}
\alias{jsondiff-package}
\title{jsondiff: render difference between JSONs}
\description{
An R interface to jsondiffpatch library \url{https://github.com/benjamine/jsondiffpatch}
}
\section{Support}{


  Use \url{https://github.com/bergant/jsondiff/issues} for bug reports
}

\examples{
\dontrun{

x1 <- list(
  name = "Pluto",
  orbital_speed_kms = 4.7,
  category = "planet",
  composition = c("methane", "nitrogen")
)

x2 <- list(
  name = "Pluto",
  category = "dwarf planet",
  orbital_speed_kms = 4.7,
  composition = c("nitrogen", "methane", "carbon monoxide")
)

jsondiff(x1, x2)

}


}
